mpox_table_update/requirements.txt
pyxform==4.0.0
elementpath>=5.0.1
pandas>=2.0
openpyxl>=3.1.3
openhexa.sdk==2.12.1
rapidfuzz>=3.9.4

